timer component refactored
